Move 'directory' submodule to own file
